Modularize & simplify db functions, Server Actions
